Don't mark targets of unconditional jumps with side effects as FALLTHRU.
When an unconditional jump with side effects targets an immediately
following label, rtl_tidy_fallthru_edge is called.  Since it has side
effects, it doesn't remove the jump, but the label is still marked
as fallthru.  This later causes a verification error.  Do nothing in this
case instead.

gcc/ChangeLog:

	* cfgrtl.c (rtl_tidy_fallthru_edge): Bail for unconditional jumps
	with side effects.