Fix NOTE_INSN_PROLOGUE_END after unconditional jump.
With the new s390 split-stack support, when optimization is enabled,
the cold path of calling __morestack is likely to be moved to the
end of the function.  This will result in the function ending in
split_stack_call_esa, which is an unconditional jump instruction and
part of the function prologue.  reposition_prologue_and_epilogue_notes
will insert NOTE_INSN_PROLOGUE_END right after it (and before the
following barrier), causing a verification error.  Insert it after
the barrier instead (and outside of basic block).

gcc/ChangeLog:

	* function.c (reposition_prologue_and_epilogue_notes): Avoid
	verification error if the last insn of prologue is an unconditional
	jump.